student and admin side frontend update

+++ database/migrations/2024_05_31_024653_parents_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('parents', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('students_id');
+            $table->unsignedBigInteger('students_id');
             $table->foreign('students_id')->references('id')->on('students')->onDelete('cascade');
             $table->string('plname');
             $table->string('pfname');
