+++ database/migrations/2024_06_10_051454_benefits_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create("benefits", function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger("programs_id");
+            $table->unsignedBigInteger("programs_id");
             $table->foreign("programs_id")->references('id')->on('programs')->onDelete('cascade');
             $table->string('benefit');
             $table->timestamps();
